+++ main/src/main/java/ru/practicum/ewm/model/compilations/dto/NewCompilationDto.java
@@ -15,7 +15,7 @@ import java.util.List;
 @NoArgsConstructor
 public class NewCompilationDto {
 
-    private List<Integer> events; //Список идентификаторов событий входящих в подборку
+    private List<Long> events; //Список идентификаторов событий входящих в подборку
 
     private Boolean pinned; //Закреплена ли подборка на главной странице сайта
