+++ main/src/main/java/ru/practicum/ewm/model/events/dto/EventRequestStatusUpdateRequest.java
@@ -13,6 +13,6 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class EventRequestStatusUpdateRequest {
-    private List<Integer> requestIds; //Идентификаторы запросов на участие в событии текущего пользователя
+    private List<Long> requestIds; //Идентификаторы запросов на участие в событии текущего пользователя
     private String status; //Новый статус запроса на участие в событии текущего пользователя
 }
