+++ main/src/main/java/ru/practicum/ewm/model/events/dto/EventShortDto.java
@@ -13,16 +13,16 @@ import javax.validation.constraints.NotNull;
 @AllArgsConstructor
 @NoArgsConstructor
 public class EventShortDto {
-    private Integer id;
+    private Long id;
     @NotBlank
     private String annotation; //Краткое описание
     @NotNull
     private CategoryDto category; //Категория
-    private Integer confirmedRequests; //Количество одобренных заявок на участие в данном событии
+    private Long confirmedRequests; //Количество одобренных заявок на участие в данном событии
     private String eventDate; //Дата и время на которые намечено событие (в формате "yyyy-MM-dd HH:mm:ss")
     private UserShortDto initiator; //Дата и время на которые намечено событие (в формате "yyyy-MM-dd HH:mm:ss")
     @NotNull
     private Boolean paid; //Нужно ли оплачивать участие
     private String title; //Заголовок
-    private Integer views; //Количество просмотрев события
+    private Long views; //Количество просмотрев события
 }
