+++ stat/stat-service/src/main/java/ru/practucum/ewm/model/EndpointHit.java
@@ -16,7 +16,7 @@ public class EndpointHit {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id; // Идентификатор записи
+    private Long id; // Идентификатор записи
     private String app; // Идентификатор сервиса для которого записывается информация
     private String uri; // URI для которого был осуществлен запрос
     private String ip; // IP-адрес пользователя, осуществившего запрос
